Move ListWrapper in hparams.
